RE-1513 Reverse release order in component files
This change reverses the expected order of releases in a component file
so they are sorted in descending order. The benefit of this change is
that it reduces the potential for merge conflicts when adding new
component releases. It is also helpful when viewing the component files
because the most recent versions are at the top.

Without this change, when adding a new major version (series), it is
placed at the bottom of the list of releases. If a new release is added
to the latest existing series, that also gets places at the bottom of
the list of releases and so, these two changes create a merge conflict.
Reversing the order ensures there is always a few lines of unmodified
text separating the two changes thereby preventing a conflict.