RE-1786 Add Python version check to component tool
The component tool will output a syntax error that is difficult to
understand when the tool is run under Python 2.x.  This change involves
modifying the rpc_component scripts to check the Python version and
instruct the user to use Python 3.2+ when the wrong version is being
used.

This change required using the "python_requires" option and the
"sys.version_info" method because using "python_requires" works well
while using the "pip install ." command but fails when manually
running "setup.py install".

JIRA: RE-1786