RE-1894 Add schema to allow RC tags
Currently only alpha/beta pre-release tags are supported.
This patch adds the schema to also accept rc pre-release
tags so that they can be used for releases.

The accepted form is the same as with alpha/beta tags, eg:
1.0.0-rc.1, 1.0.0-rc.2, etc.

Due to the increase in scope, the capture group name
a_or_b and related maps are all renamed to use the more
generic term prerelease.