RE-1599 Add a dependents lookup option
This task requires a reverse dependency lookup option, or in other words
an option that looks up components that are dependent on the specified
component.  The functionality will mainly be used by common.groovy in
order to find which repos need to be dep updated, as further described
in RE-1599.  In addition, this change also includes recommended tweaks
from the review process and a rename of "gating/pre_merge_test" to
"gating/check" as a fix that is described in RE-1905.

JIRA: RE-1599